+++ tests/unit/TmLiStakeTransaction.spec.js
@@ -12,6 +12,12 @@ describe("TmLiStakeTransaction", () => {
         description: {
           moniker: "cool validator"
         }
+      },
+      {
+        owner: "cosmosvaladdr157mg9hnhchfrqvk3enrvmvj29yhmlwf759xrgw",
+        description: {
+          moniker: "Kentucky val"
+        }
       }
     ],
     validatorURL: "/validator"